Do not call exit()
Using exit() in a shared library is not considered a good practice, at
least by Lintian, the static analysis tool for Debian packages:
“In the case of an error, the library should instead return an
appropriate error code to the calling program which can then determine
how to handle the error, including performing any required clean-up.”
(see https://lintian.debian.org/tags/exit-in-shared-library)

This commit replaces the calls to exit() by the setting of a global
variable (vc_last_error_message) with the string describing the error
and appropriate return values for the concerned functions. This
variable is not exported as a symbol in the shared library. Hence, two
new functions are created for setting/getting its value:
vc_set_error_message and vc_get_error_message. Only the former is
exported as a symbol of the shared library.

This implies a change in ABI, with the addition of
vc_get_error_message. There are no other changes in the ABI besides
that.

Visibility of the symbols is controlled (when using GNU CC) by a
technique described in
https://www.gnu.org/software/gnulib/manual/html_node/Exported-Symbols-of-Shared-Libraries.html

As regards the flex code, exit is colled in function yy_fatal_error,
which is default definition of the YY_FATAL_ERROR macro. Since
YY_FATAL_ERROR() is called all over the place, it is not possible to
use yy_fatal_error. A solution using setjmp/longjmp was implemented,
with the jump buffer being set in the function parse_vcard_file,
around the call to yyparse().

Specific changes:

* configure.ac:
  + Change LIBVC_ABI_VERSION to 1:2:1
  + Include m4/visibility.m4 and call gl_VISIBILITY

* m4/visibility.m4: New file

* doc/vc.3: Add vc_get_error_message to the list of API functions

* src/vc.c:
  + Create variable vc_last_error_message
  + Add function vc_set_error_message
  + Add function vc_get_error_message
  + In functions vc_new and vc_param_new, replace the calls to exit by
    calls to vc_set_error_message

* src/vc_parse.y:
  + Define jump buffer scan_recovery
  + Surrond the call to yyparse() with setjmp()

* src/vc_scan.l:
  + Define YY_FATAL_ERROR as vc_flex_fatal_error
  + Add function vc_flex_fatal_error, which makes a long jump to
    scan_recovery